+++ src/compositions/useRoute.ts
@@ -1,10 +1,10 @@
 import { watch } from 'vue'
 import { useRouter } from '@/compositions/useRouter'
 import { UseRouteInvalidError } from '@/errors'
+import { isRoute } from '@/guards/routes'
 import { RouterRoute, isRouterRoute } from '@/services/createRouterRoute'
 import { RegisteredRouteMap } from '@/types/register'
 import { ResolvedRoute } from '@/types/resolved'
-import { isRoute } from '@/utilities/routes'
 
 export type UseRouteOptions = {
   exact?: boolean,
