+++ src/main.ts
@@ -3,4 +3,4 @@ export * from './compositions'
 export * from './errors'
 export * from './services'
 export * from './types'
-export { isRoute } from './utilities/routes'
+export { isRoute } from './guards/routes'
